Completed the Input with Validation example
